virt: acrn: support asyncio in ioeventfd
In ACRN, we use the ioeventfd in vhost device. We also support the
virtio device using ioveventfd for FE to NOTIFY BE. But in current
implementation, when the User VM write the NOTIFY register to notify BE,
it will trap to hypervisor. After the write access is finished, User VM
can return.
To return earlier, we provide an asyncio mode for an I/O request. For
asyncio, hypervisor directly put the request in an asyncio shared buffer,
and then return to User VM. To take use of this mechanism, add a new flag
ACRN_IOEVENTFD_FLAG_ASYNCIO to indicate that the IO is asyncio. And use
hypercall to register the asyncio info to hypervisor.

Signed-off-by: Conghui <conghui.chen@intel.com>
Reviewed-by: Fei Li <fei1.li@intel.com>