virt: acrn: remove WQ_UNBOUND flag
Due to the two level schedule model in type 1 hypervisor, Service OS
kernel's scheduler does not aware the running status of the User VM,
which means, a thread may be schedule to a CPU which is idle in Service
OS but actually running in User VM.
For the ioreq dispatch workqueue, as it was in the critical path of I/O
request, and ACRN hypervisor only inject the ioreq interrupt to CPU0, so
migrate to other CPU is meaningless and will bring more latency.
So, remove the WQ_UNBOUND flag, and let CPU0 do the dispatch directly.

Signed-off-by: Conghui <conghui.chen@intel.com>
Reviewed-by: Fei Li <fei1.li@intel.com>