virt: acrn: add shared buffer interface
In ACRN, the shared buffer is created to share data between hypervisor,
kernel and device model (in userpace). This buffer can be allocated and
initialized by kernel or device model.

Add hypercall to register shared buffer(sbuf) to ACRN hyperviosr. The
'vm_id' is used to indicate who owns this sbuf. The 'buffer' is the base
address for sbuf.

Signed-off-by: Conghui <conghui.chen@intel.com>
Reviewed-by: Fei Li <fei1.li@intel.com>